Refactors the _app and criates a ContextProvider
