+++ src/pages/_app.tsx
@@ -2,30 +2,11 @@ import "../styles/global.scss";
 import { Header } from "../components/Header";
 import { Player } from "../components/Player";
 import styles from "../styles/app.module.scss";
-import { PlayerContext } from "../contexts/PlayerContext";
-import { useState } from "react";
+import { PlayerContextProvider } from "../contexts/PlayerContext";
 
 function MyApp({ Component, pageProps }) {
-    const [episodes, setEpisodes] = useState([]);
-    const [currentEpisodeIndex, setCurrentEpisodeIndex] = useState(0);
-    const [isPlaying, setIsPlaying] = useState(false);
-
-    function play(episode) {
-        setEpisodes([episode]);
-        setCurrentEpisodeIndex(0);
-        setIsPlaying(true);
-    }
-
-    function togglePlay() {
-        setIsPlaying(!isPlaying);
-    }
-
-    function setPlayingState(state: boolean) {
-        setIsPlaying(state);
-    }
-
     return (
-        <PlayerContext.Provider value={{ episodes, currentEpisodeIndex, isPlaying, play, togglePlay, setPlayingState }}>
+        <PlayerContextProvider>
             <div className={styles.wrapper}>
                 <main>
                     <Header />
@@ -33,7 +14,7 @@ function MyApp({ Component, pageProps }) {
                 </main>
                 <Player />
             </div>
-        </PlayerContext.Provider>
+        </PlayerContextProvider>
     );
 }
